% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_step_dist.R
\name{calc_dist_OU_backward}
\alias{calc_dist_OU_backward}
\title{Calculate reverse-time OU distribution.}
\usage{
calc_dist_OU_backward(num_states, fixed_time, fixed_speed, times, behavs,
  mean_para, corr_para, var_para)
}
\arguments{
\item{num_states}{Numeric, number of behavioural states.}

\item{fixed_time}{Numeric, of the time at the end of the section.}

\item{fixed_speed}{Numeric, of the speed at the end of the section.}

\item{times}{Vector of refined times.}

\item{behavs}{Vector of refined behaviours.}

\item{mean_para}{Vector of mean speed parameters.}

\item{corr_para}{Vector of speed correlation parameters.}

\item{var_para}{Vector of speed variance parameters.}
}
\value{
List with the components:
  \describe{
  \item{\code{mean}}{Vector of mean.}
  \item{\code{covar}}{Square matrix of covariances.}
  }
}
\description{
\code{calc_dist_OU_backward} returns the distribution of a reverse time OU process.
}
\seealso{
Other Step distribution: \code{\link{calc_dist_OU_bridge}},
  \code{\link{calc_dist_OU_forward}},
  \code{\link{calc_dist_step_backward}},
  \code{\link{calc_dist_step_bridge}},
  \code{\link{calc_dist_step_forward}},
  \code{\link{calc_step_dist}}
}
\concept{Step distribution}
