% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_step_dist.R
\name{calc_dist_OU_bridge}
\alias{calc_dist_OU_bridge}
\title{Calculate OU bridge distribution.}
\usage{
calc_dist_OU_bridge(num_states, fixed_times, fixed_behav, fixed_speeds, times,
  behavs, mean_para, corr_para, var_para)
}
\arguments{
\item{num_states}{Numeric, number of behavioural states.}

\item{fixed_times}{Vector, length 2, of the times at the ends of the section.}

\item{fixed_behav}{Numeric, fixed behaviour at the start.}

\item{fixed_speeds}{Vector, length 2, of the speeds at the ends of the section.}

\item{times}{Vector of refined times.}

\item{behavs}{Vector of refined behaviours.}

\item{mean_para}{Vector of mean speed parameters.}

\item{corr_para}{Vector of speed correlation parameters.}

\item{var_para}{Vector of speed variance parameters.}
}
\value{
List with the components:
  \describe{
  \item{\code{mean}}{Vector of mean.}
  \item{\code{covar}}{Square matrix of covariances.}
  }
}
\description{
\code{calc_dist_OU_bridge} returns the distribution of an OU bridge.
}
\seealso{
Other Step distribution: \code{\link{calc_dist_OU_backward}},
  \code{\link{calc_dist_OU_forward}},
  \code{\link{calc_dist_step_backward}},
  \code{\link{calc_dist_step_bridge}},
  \code{\link{calc_dist_step_forward}},
  \code{\link{calc_step_dist}}
}
\concept{Step distribution}
