% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_step_dist.R
\name{calc_dist_step_forward}
\alias{calc_dist_step_forward}
\title{Step distribution of a path section (correlated steps, end section).}
\usage{
calc_dist_step_forward(fixed_constant, fixed_values, times, behavs, move_params)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{num_states}}{Numeric, number of behavioural states.}
\item{\code{q.speed_mean}}{Vector, index values in \code{move_params} which correspond to mean speed parameters.}
\item{\code{q.speed_corr}}{Vector, index values in \code{move_params} which correspond to speed correlation parameters.}
\item{\code{q.speed_var}}{Vector, index values in \code{move_params} which correspond to speed variance parameters.}
}}

\item{fixed_values}{List with components:
\describe{
\item{\code{time}}{Numeric, the time at the start of the section.}
\item{\code{after_time}}{Numeric, the time after the end point (used to work out the end speed).}
\item{\code{behav}}{Numeric, fixed behaviour at the start.}
\item{\code{step}}{Numeric, the step at the start of the section.}
}}

\item{times}{Vector of refined times.}

\item{behavs}{Vector of refined behaviours.}

\item{move_params}{Vector of movement parameters.}
}
\value{
List with the components:
  \describe{
  \item{\code{mean}}{Vector of mean step lengths.}
  \item{\code{covar}}{Square matrix of step length covariances.}
  }
}
\description{
\code{calc_dist_step_forward} returns the distribution of a set of steps when the correlated model is assumed and the section at the end of the full path.
}
\seealso{
Other Step distribution: \code{\link{calc_dist_OU_backward}},
  \code{\link{calc_dist_OU_bridge}},
  \code{\link{calc_dist_OU_forward}},
  \code{\link{calc_dist_step_backward}},
  \code{\link{calc_dist_step_bridge}},
  \code{\link{calc_step_dist}}
}
\concept{Step distribution}
