% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_loc_dist.R
\name{calc_dist_true_locs_forward}
\alias{calc_dist_true_locs_forward}
\title{Calculate distribution of fixed locations.}
\usage{
calc_dist_true_locs_forward(start_loc, fixed_index, bearings, step_dist)
}
\arguments{
\item{start_loc}{Vector (length 2) of the x,y starting position of the section to update.}

\item{fixed_index}{Vector (variable length, at least 1) of the index points on the section where a fixed location is (will always include the last point).}

\item{bearings}{Vector (variable length) of the actual bearings on the section.}

\item{step_dist}{List with components:
\describe{
\item{\code{mean}}{Vector of the mean step lengths on the section.}
\item{\code{covar}}{Square matrix of the covariances between the steps on the section.}
}}
}
\value{
List with components:
  \describe{
  \item{\code{A}}{Linear constraint matrix.}
  \item{\code{loc_mean}}{Vector of location means (arranged as x_1,y_1,x_2,...).}
  \item{\code{loc_covar}}{Square matrix of location covariances.}
  \item{\code{loc_step_covar}}{Matrix of covariances between the steps and the locations.}
  }
}
\description{
\code{calc_dist_true_locs_forward} returns the distribution of a set of locations, given step distribution and actual bearings.
}
\seealso{
Other Location distribution: \code{\link{calc_dist_true_locs_backward}}
}
\concept{Location distribution}
