% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calc_error_lik}
\alias{calc_error_lik}
\title{Calculate observation error (correlated) likelihood}
\usage{
calc_error_lik(obs, error_params)
}
\arguments{
\item{obs}{List with componenets:
\describe{
\item{\code{errors}}{List with componenets:
  \describe{
  \item{\code{x}}{Vector of errors in x direction.}
  \item{\code{y}}{Vector of errors in y direction.}
  }}
\item{\code{time_diffs}}{Vector of differences in the observation time scale.}
}}

\item{error_params}{Vector (length 2) of the error variance and correlation.}
}
\value{
Numeric, likelihood.
}
\description{
\code{calc_error_lik} returns the likelihood of a set of errors (correlated).
}
\seealso{
Other Observation error parameters: \code{\link{Gibbs_update_obs_error_param}},
  \code{\link{MH_update_obs_error_param}}
}
\concept{Observation error parameters}
