% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calc_markov_steady_state}
\alias{calc_markov_steady_state}
\title{Markov chain equilibrium distribution.}
\usage{
calc_markov_steady_state(num_states, behav_params)
}
\arguments{
\item{num_states}{Numeric, number of behavioural states.}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector (length \code{num_states}) of switching rates.}
\item{\code{q}}{Matrix (square size \code{num_states} with \code{NA} diagonal elements) of switching probabilities. Note: row = from, column = to.}
}}
}
\value{
List with components:
  \describe{
  \item{\code{steady_state}}{Vector (length \code{num_states}) giving the equilibrium distribution.}
  \item{\code{generator}}{Matrix (square, size \code{num_states}) giving the generator matrix of the Markov chain (just an alternative presentation of behav params).}
  }
}
\description{
\code{calc_markov_steady_state} returns the equilibrium distribution of a Markov chain.
}
\examples{
para <- list(lambda = c(0.1, 0.2),
             q = matrix(c(NA, 1, 1, NA), nrow = 2))
calc_markov_steady_state(2, para)
# $steady_state
# [1] 0.6666667 0.3333333
#
# $generator
# [,1] [,2]
# [1,] -0.1  0.1
# [2,]  0.2 -0.2

para <- list(lambda = c(0.1, 0.2, 0.3),
             q = matrix(c(NA, 0.5, 0.2, 0.3, NA, 0.8, 0.7, 0.5, NA), nrow = 3))
calc_markov_steady_state(3, para)
$steady_state
# [1] 0.4568528 0.3274112 0.2157360
#
# $generator
# [,1]  [,2]  [,3]
# [1,] -0.10  0.03  0.07
# [2,]  0.10 -0.20  0.10
# [3,]  0.06  0.24 -0.30
}
\seealso{
Other Speed parameters: \code{\link{calc_cond_step_dist}},
  \code{\link{calc_step_lik}},
  \code{\link{update_speed_param}}

Other Markov chain simulations: \code{\link{sim_markov_backward}},
  \code{\link{sim_markov_bridge}},
  \code{\link{sim_markov_forward}}
}
\concept{Markov chain simulations}
\concept{Speed parameters}
