% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calc_markov_suff_stats}
\alias{calc_markov_suff_stats}
\title{Calculate Markov chain sufficient statistics.}
\usage{
calc_markov_suff_stats(num_states, switch_times, switch_states)
}
\arguments{
\item{num_states}{Numeric, number of behavioural states.}

\item{switch_times}{Vector, times of behavioural switches.}

\item{switch_states}{Vector, states at switching times.}
}
\value{
List with components:
  \describe{
  \item{\code{time_in_state}}{Vector (length \code{num_states}) with total time spent in each behavioural state.}
  \item{\code{num_trans}}{Matrix (square, size \code{num_states} with \code{NA} diagonal elements) with number of transitions from each state (row) to every other (column).}
  }
}
\description{
\code{calc_markov_suff_stats} calculates the sufficient statistics (time in each state and the number of transitions between state pairs) for a Markov chain.
}
\examples{
times <- c(0, 12, 15, 20, 26, 43, 50)
states <- c(1, 2, 1, 2, 1, 2, 2)

calc_markov_suff_stats(2, times, states)
# $time_in_state
# [1] 34 16
#
# $num_trans
# [,1] [,2]
# [1,]   NA    3
# [2,]    2   NA

times <- c(100, 140, 167, 189, 241, 264, 279, 304, 365, 397, 400)
states <- c(1, 2, 1, 3, 2, 3, 1, 2, 3, 2, 2)
calc_markov_suff_stats(3, times, states)
# time_in_state
# [1]  87 114  99
#
# $num_trans
# [,1] [,2] [,3]
# [1,]   NA    2    1
# [2,]    1   NA    2
# [3,]    1    2   NA
}
\seealso{
Other Behaviour parameters: \code{\link{calc_markov_post_param}},
  \code{\link{check_valid_markov}},
  \code{\link{update_behav_param}}
}
\concept{Behaviour parameters}
