% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_markov.R
\name{calc_reverse_markov_gen}
\alias{calc_reverse_markov_gen}
\title{Calculate the generator matrix of a Markov chain in reverse time.}
\usage{
calc_reverse_markov_gen(num_states, behav_params)
}
\arguments{
\item{num_states}{Numeric, number of (behavioural) states in the Markov chain.}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector of length \code{num_states}, switching rates out of each state.}
\item{\code{q}}{Sqaure matrix size \code{num_states}, probability of switching from (row) each state to (column) another. Diagonal elements are NA.}
}}
}
\value{
List with the components:
  \describe{
  \item{\code{times}}{Vector, switching times of the simulation. Includes the fixed end points.}
  \item{\code{states}}{Vector, states at switching times of the simulation. Includes the fixed end points.}
  }
}
\description{
\code{calc_reverse_markov_gen} returns the generator matrix of the Markov chain which is the time-reversed process of a given Markov chain.
}
\examples{
# In a two state process, the reverse chain is equal to the forward chain
num_states <- 2
behav_params <- list(lambda = c(0.1,0.2),
                     q = matrix(c(NA,1,1,NA), nrow = 2))

calc_reverse_markov_gen(num_states, behav_params)
# $lambda
# [1] 0.1 0.2
# 
# $q
# [,1] [,2]
# [1,]   NA    1
# [2,]    1   NA


num_states <- 3
behav_params <- list(lambda = c(0.1,0.2,0.3),
                     q = matrix(c(NA,0.15/0.2,0.05/0.3,0.05/0.1,NA,0.25/0.3,0.05/0.1,0.05/0.2,NA),nrow=3))

calc_reverse_markov_gen(num_states, behav_params)
# $lambda
# [1] 0.1 0.2 0.3
# 
# $q
# [,1]      [,2]      [,3]
# [1,]        NA 0.8684211 0.1315789
# [2,] 0.4318182        NA 0.5681818
# [3,] 0.6333333 0.3666667        NA
}
\seealso{
\code{\link{calc_markov_steady_state}} for calculating a Markov steady state/equilibrium distribution and
 \code{\link{sim_markov_backward}} for simulating from this reverse-time process.
}
