% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{calc_step_lik}
\alias{calc_step_lik}
\title{Step likelihood.}
\usage{
calc_step_lik(fixed_constant, refined_path, move_params, behav_params)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{num_states}}{Numeric, number of behavioural states.}
\item{\code{indep_step}}{Boolean, True if model is independent step, False if it is correlated step.}
\item{\code{q.speed_mean}}{Vector, index positions within \code{move_params} relating to speed mean parameters.}
\item{\code{q.speed_corr}}{Vector, index positions within \code{move_params} relating to speed correlation parameters.}
\item{\code{q.speed_var}}{Vector, index positions within \code{move_params} relating to speed variance parameters.}
}}

\item{refined_path}{List with components:
\describe{
\item{\code{time_diffs}}{Vector, differences in the refined time scale.}
\item{\code{behavs}}{Vector, refined behavioural/state sequence.}
\item{\code{steps}}{Vector, refined step process.}
}}

\item{move_params}{Vector of movement parameters (as turn_var_1, turn_var_2, mean_speed_1, .., speed_var_2).}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector (length \code{num_states}) of switching rates.}
\item{\code{q}}{Matrix (square size \code{num_states} with \code{NA} diagonal elements) of switching probabilities. Note: row = from, column = to.}
}}
}
\value{
Numeric, likelihood of the set of steps.
}
\description{
\code{calc_step_lik} returns the likelihood of a set of steps given speed parameters.
}
\examples{
const <- list(num_states = 3,
              indep_step = FALSE,
              q.speed_mean = 4:6,
              q.speed_corr = 7:9,
              q.speed_var = 10:12)
path <- list(steps = c(5.3, 6.8, 15.3, 14.5, 9.4, 3.4, 8.5, 14.6, 17.8, 22.4, 16.7),
             time_diffs = c(1, 1, 1.5, 1.5, 1, 0.5, 1, 1, 1, 2, 1),
             behavs = c(1, 1, 1, 2, 2, 3, 3, 2, 2, 2, 1))
move_params <- c(NA, NA, NA, 5, 10, 3, 0.1, 0.2, 0.3, 4, 6, 2)
behav_params <- list(lambda = c(0.1, 0.2, 0.3),
                     q = matrix(c(NA, 0.5, 0.2, 0.3, NA, 0.8, 0.7, 0.5, NA), nrow = 3))

calc_step_lik(const, path, move_params, behav_params)
# [1] -78.24704

path <- list(steps = c(5.3, 6.8, 15.3, 14.5, 9.4, 3.4, 8.5, 14.6, 17.8, 22.4, 16.7),
             time_diffs = c(2, 2, 0.5, 1, 2, 0.25, 2, 1.5, 1, 1.2, 0.6),
             behavs = c(1, 1, 1, 2, 2, 3, 3, 2, 2, 2, 1))
calc_step_lik(const, path, move_params, behav_params)
# [1] -791.1856
}
\seealso{
Other Speed parameters: \code{\link{calc_cond_step_dist}},
  \code{\link{calc_markov_steady_state}},
  \code{\link{update_speed_param}}
}
\concept{Speed parameters}
