% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{check_valid_markov}
\alias{check_valid_markov}
\title{Check Markov chain is valid.}
\usage{
check_valid_markov(num_states, switch_times, switch_states)
}
\arguments{
\item{num_states}{Numeric, number of behavioural states.}

\item{switch_times}{Vector, times of behavioural switches.}

\item{switch_states}{Vector, states at switching times.}
}
\value{
\code{NULL}
}
\description{
\code{check_valid_markov} is a testing function to check that a Markov chain is valid.
}
\seealso{
Other Behaviour parameters: \code{\link{calc_markov_post_param}},
  \code{\link{calc_markov_suff_stats}},
  \code{\link{update_behav_param}}
}
\concept{Behaviour parameters}
