% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_path_section.R
\name{create_behav_proc_from_refined}
\alias{create_behav_proc_from_refined}
\title{Create behavioural switch times and states.}
\usage{
create_behav_proc_from_refined(times, states)
}
\arguments{
\item{times}{Vector of switch times.}

\item{states}{Vector of switch states.}
}
\value{
List with the components:
  \describe{
  \item{\code{times}}{Vector, switching times of the simulation.}
  \item{\code{states}}{Vector, states at switching times of the simulation.}
  }
}
\description{
\code{create_behav_proc_from_refined} creates behavioural process as switch times and states given refined behaviours.
}
\examples{
times <- c(0.000000, 0.460000, 0.920000, 1.380000, 1.840000, 2.300000,
           2.787368, 3.274737, 3.762105, 4.249474, 4.736842, 5.224211,
           5.711579, 6.198947, 6.686316, 7.173684, 7.661053, 8.148421,
           8.635789, 9.123158, 9.610526, 10.097895, 10.585263, 11.072632,
           11.560000, 12.051429, 12.542857, 13.034286, 13.525714, 14.017143,
           14.508571, 15.000000, 15.433333, 15.866667, 16.300000, 16.733333,
           17.166667, 17.600000, 18.080000, 18.560000, 19.040000, 19.520000,
           20.000000)
states <- c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
            2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 2, 2, 2, 2, 2, 2)

create_behav_proc_from_refined(times, states)
# $times
# [1]  0.00  2.30 11.56 17.60 20.00
#
# $states
# [1] 1 2 3 2 2
}
\seealso{
Other Refined scales: \code{\link{create_refined_behavs}},
  \code{\link{create_refined_times}}
}
\concept{Refined scales}
