% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_path_section.R
\name{draw_random_path_section}
\alias{draw_random_path_section}
\title{Draw a random path section.}
\usage{
draw_random_path_section(section_length_limits, obs_index)
}
\arguments{
\item{section_length_limits}{Vector of length two, lower and upper length limits for the section to choose.}

\item{obs_index}{Vector of length number of observations, giving the index on the refined path where observation times coincide.}
}
\value{
List with components:
  \describe{
  \item{\code{start}}{Numeric, the index of the start of chosen section.}
  \item{\code{end}}{Numeric, the index of the end of chosen section.}
  \item{\code{contained_obs_index}}{Vector (length variable) of index which are also observations (within the section).}
  \item{\code{contained_obs}}{Vector (length variable) of observation indexs included in the section.}
  }
}
\description{
\code{draw_random_path_section} returns details on a randomly selected section of path to update.
}
\examples{
limits <- c(3, 10)
obs_index <- c(0,5,10,18,24,28,40,45,51,56,70)

set.seed(1234)
draw_random_path_section(limits, obs_index)
# $start
# [1] 10
#
# $end
# [1] 15
#
# $contained_obs_index
# numeric(0)
#
# $contained_obs
# integer(0)

set.seed(7654)
draw_random_path_section(limits, obs_index)
# $start
# [1] 46
#
# $end
# [1] 54
#
# $contained_obs_index
# [1] 51
#
# $contained_obs
# [1] 9

limits <- c(8, 15)
set.seed(678)
draw_random_path_section(limits, obs_index)
# $start
# [1] 14
#
# $end
# [1] 27
#
# $contained_obs_index
# [1] 18 24
#
# $contained_obs
# [1] 4 5
}
\seealso{
Other Choose path: \code{\link{set_current_values}},
  \code{\link{set_fixed_values}}
}
\concept{Choose path}
