% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{Gibbs_update_obs_error_param}
\alias{Gibbs_update_obs_error_param}
\title{Simulate observation error (independent) parameter.}
\usage{
Gibbs_update_obs_error_param(error_prec_prior, refined_path, obs)
}
\arguments{
\item{error_prec_prior}{List with components:
\describe{
  \item{\code{rate}}{Numeric with prior rate of the error precision (not variance).}
  \item{\code{shape}}{Numeric with prior shape of the error precision (not variance).}
}}

\item{refined_path}{List with components:
\describe{
\item{\code{X}}{Vector of the refined x locations on the path.}
\item{\code{Y}}{Vector of refined y locations on the path.}
}}

\item{obs}{List with componenets:
\describe{
\item{\code{x}}{Observed x locations.}
\item{\code{y}}{Observed y locations.}
\item{\code{index_on_refined_path}}{Index values on the refined path where observations occur.}
}}
}
\value{
Numeric, sampled error variance.
}
\description{
\code{Gibbs_update_obs_error_param} returns a conditional Gibbs sample of the observation error parameter when the errors are assumed independent.
}
\seealso{
Other Observation error parameters: \code{\link{MH_update_obs_error_param}},
  \code{\link{calc_error_lik}}
}
\concept{Observation error parameters}
