% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{MH_update_obs_error_param}
\alias{MH_update_obs_error_param}
\title{Simulate observation error (correlated) parameters.}
\usage{
MH_update_obs_error_param(fixed_constant, obs, curr_move_params)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
  \item{\code{q.obs_error}}{Vector with index positions in \code{move_params} that correspond to the error parameters.}
  \item{\code{error_pert_sd}}{Vector (length 2) of perturbation standard deviations of the error parameters.}
}}

\item{obs}{List with componenets:
\describe{
\item{\code{errors}}{List with componenets:
  \describe{
  \item{\code{x}}{Vector of errors in x direction.}
  \item{\code{y}}{Vector of errors in y direction.}
  }}
\item{\code{time_diffs}}{Vector of differences in the observation time scale.}
}}

\item{curr_move_params}{Vector of parameters.}
}
\value{
Numeric, sampled error variance.
}
\description{
\code{MH_update_obs_error_param} returns an MH sample of the observation error parameters when the errors are assumed correlated.
}
\seealso{
Other Observation error parameters: \code{\link{Gibbs_update_obs_error_param}},
  \code{\link{calc_error_lik}}
}
\concept{Observation error parameters}
