% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_path_section.R
\name{set_current_values}
\alias{set_current_values}
\title{Identify current values relating to a chosen path section.}
\usage{
set_current_values(type, drawn_section, refined_path)
}
\arguments{
\item{type}{Character, whether the path section is at the "start, "end", or "middle" of the full path.}

\item{drawn_section}{List, output from \code{draw_random_path_section}.}

\item{refined_path}{List with components:
\describe{
\item{\code{times}}{Vector, refined time scale.}
\item{\code{behavs}}{Vector, refined behavioural/state sequence.}
\item{\code{bearings}}{Vector, refined bearing process.}
\item{\code{steps}}{Vector, refined step process.}
\item{\code{X}}{Vector, refined x location.}
\item{\code{Y}}{Vector, refined y location.}
}}
}
\value{
List of current values.
}
\description{
\code{set_current_values} returns details on the current values to be simulated.
}
\seealso{
Other Choose path: \code{\link{draw_random_path_section}},
  \code{\link{set_fixed_values}}
}
\concept{Choose path}
