% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_path_section.R
\name{set_fixed_values}
\alias{set_fixed_values}
\title{Identify fixed values relating to a chosen path section.}
\usage{
set_fixed_values(type, drawn_section, refined_path, obs)
}
\arguments{
\item{type}{Character, whether the path section is at the "start, "end", or "middle" of the full path.}

\item{drawn_section}{List, output from \code{draw_random_path_section}.}

\item{refined_path}{List with components:
\describe{
\item{\code{times}}{Vector, refined time scale.}
\item{\code{behavs}}{Vector, refined behavioural/state sequence.}
\item{\code{bearings}}{Vector, refined bearing process.}
\item{\code{steps}}{Vector, refined step process.}
\item{\code{X}}{Vector, refined x location.}
\item{\code{Y}}{Vector, refined y location.}
}}

\item{obs}{List with components:
\describe{
\item{\code{times}}{Vector, observation time scale.}
\item{\code{x}}{Vector, observed x location.}
\item{\code{y}}{Vector, observed y location.}
}}
}
\value{
List of fixed values.
}
\description{
\code{set_fixed_values} returns details on the fixed values needed to update a path section.
}
\seealso{
Other Choose path: \code{\link{draw_random_path_section}},
  \code{\link{set_current_values}}
}
\concept{Choose path}
