% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_markov.R
\name{sim_markov_backward}
\alias{sim_markov_backward}
\title{Simulate Markov chain backwards in time.}
\usage{
sim_markov_backward(num_states, fixed_values, behav_params)
}
\arguments{
\item{num_states}{Numeric, number of (behavioural) states in the Markov chain.}

\item{fixed_values}{List with components:
\describe{
\item{\code{inc_times}}{Vector of length two, start and end times of the simulation.}
\item{\code{inc_behav}}{Vector of length one, end known value (behaviour) of the simulation.}
}}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector of length \code{num_states}, switching rates out of each state.}
\item{\code{q}}{Sqaure matrix size \code{num_states}, probability of switching from (row) each state to (column) another. Diagonal elements are NA.}
}}
}
\value{
List with the components:
  \describe{
  \item{\code{times}}{Vector, switching times of the simulation. Includes the fixed end points.}
  \item{\code{states}}{Vector, states at switching times of the simulation. Includes the fixed end points.}
  }
}
\description{
\code{sim_markov_backward} returns a simulated realisation of a Markov chain backwards in time from a known value and time.
}
\examples{
end_points <- list(inc_times = c(0,10),
                   inc_behav = c(3))
param <- list(lambda = c(0.1,0.2,0.3),
              q = matrix(c(NA,0.5,0.3,0.35,NA,0.7,0.65,0.5,NA),nrow=3))

set.seed(123)
sim_markov_backward(3, end_points, param)
# $times
# [1]  0.000000  7.188476 10.000000
#
# $states
# [1] 1 3 3

set.seed(2712)
sim_markov_backward(3, end_points, param)
# $times
# [1]  0.000000  4.467302  4.691276  8.101915 10.000000
#
# $states
# [1] 2 3 1 3 3
}
\seealso{
\code{\link{calc_markov_steady_state}} and \code{\link{calc_reverse_markov_gen}} for calculating
  Markov steady state/equilibrium distribution and reverse-time generator matrices, respectively.

Other Markov chain simulations: \code{\link{calc_markov_steady_state}},
  \code{\link{sim_markov_bridge}},
  \code{\link{sim_markov_forward}}
}
\concept{Markov chain simulations}
