% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refined_path.R
\name{sim_steps}
\alias{sim_steps}
\title{Simulate conditional steps.}
\usage{
sim_steps(step_mean, step_covar, A, loc_covar_error, error_mean, error_covar,
  loc_step_covar, start_loc, fixed_locs, type)
}
\arguments{
\item{step_mean}{Vector, mean of unconditional steps.}

\item{step_covar}{Square matrix, covariance of unconditional steps.}

\item{A}{linear constraint matrix.}

\item{loc_covar_error}{Square matrix, covariance of observed locations (including error if applicable).}

\item{error_mean}{Vector, mean of the errors.}

\item{error_covar}{Square matrix, covariance of the errors.}

\item{loc_step_covar}{Matrix, covariance between unconditional steps and the observed locations.}

\item{start_loc}{Vector, length 2, giving the location at the start of the section.}

\item{fixed_locs}{Vector, giving the observed locations along the section.}

\item{type}{Character ("start", "middle", "end") describing where on the full path the section is from.}
}
\value{
Vector of simulated steps.
}
\description{
\code{sim_steps} simulates a set of conditional steps.
}
\seealso{
Other Refined path: \code{\link{update_refined_path}}
}
\concept{Refined path}
