% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{update_bearing_param}
\alias{update_bearing_param}
\title{Simulate bearing parameters.}
\usage{
update_bearing_param(fixed_constant, refined_path)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{num_states}}{Numeric, number of behavioural states.}
\item{\code{turn_prec_prior}}{List with components:
  \describe{
  \item{\code{rate}}{Vector (length \code{num_states}) with prior rate of the turn precision (not variance).}
  \item{\code{shape}}{Vector (length \code{num_states}) with prior shape of the turn precision (not variance).}
  }}
}}

\item{refined_path}{List with components:
\describe{
\item{\code{times}}{Vector of the refined times on the path.}
\item{\code{behavs}}{Vector of refined behaviours.}
\item{\code{bearings}}{Vector of refined bearings}
}}
}
\value{
Vector, sampled bearing variances for each behavioural state.
}
\description{
\code{update_bearing_param} returns a conditional Gibbs sample of the bearing parameters.
}
\examples{
const <- list(num_states = 3,
              turn_prec_prior = list(rate = c(1, 2, 3),
                                     shape = c(0.1, 0.2, 0.3)))
path <- list(times = 1:20,
             behavs = c(1, 1, 1, 1, 2, 2, 2, 3, 1, 1, 3, 3, 3, 2, 2, 2, 2, 2, 2),
             bearings = cumsum(rnorm(19, 0, 1)))

set.seed(1234)
update_bearing_param(const, path)
# [1] 4.8920033 0.8789057 1.7115489

set.seed(765)
update_bearing_param(const, path)
# [1] 2.0455226 0.7477887 2.6614361
}
\concept{Bearing parameters}
