% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{update_behav_param}
\alias{update_behav_param}
\title{Simulate behaviour parameters.}
\usage{
update_behav_param(fixed_constant, behav_proc)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{num_states}}{Numeric, number of behavioural states.}
\item{\code{prior_param}}{List with components:
  \describe{
  \item{\code{lambda_rate}}{Vector (length \code{num_states}) with prior rate of the switching rates.}
  \item{\code{lambda_shape}}{Vector (length \code{num_states}) with prior shape of the switching rates.}
  \item{\code{q_conc}}{Matrix (square, size \code{num_states} with \code{NA} diagonal elements) with prior concentration of the switching probabilities.}
  }}
}}

\item{behav_proc}{List with components:
\describe{
\item{\code{times}}{Vector, switch times of the behaviour process.}
\item{\code{states}}{Vector, behavioural states at switching times.}
}}
}
\value{
List with the components:
  \describe{
  \item{\code{lambda}}{Vector (length \code{num_states}) of sampled switching rates.}
  \item{\code{q}}{Matrix (square size \code{num_states} with \code{NA} diagonal elements) of sampled switching probabilities. Note: row = from, column = to.}
  }
}
\description{
\code{update_behav_param} returns a conditional Gibbs sample of the behaviour parameters.
}
\examples{
const <- list(prior_param = list(lambda_rate = c(0.1, 0.2),
                                 lambda_shape = c(2, 1),
                                 q_conc = matrix(c(NA, 1, 1, NA), nrow = 2)),
              num_states = 2)
behav_proc <- list(times = c(0, 12, 15, 20, 26, 43, 50),
                   states = c(1, 2, 1, 2, 1, 2, 2))

set.seed(1234)
update_behav_param(const, behav_proc)
# $lambda
# [1] 0.06755656 0.18652884
#
# $q
# [,1] [,2]
# [1,]   NA    1
# [2,]    1   NA

const <- list(prior_param = list(lambda_rate = c(0.1, 0.2, 0.3),
                                 lambda_shape = c(4, 3, 2),
                                 q_conc = matrix(c(NA, 0.5, 0.3, 0.2, NA, 0.7, 0.8, 0.5, NA), nrow = 3)),
              num_states = 3)
behav_proc <- list(times = c(100, 140, 167, 189, 241, 264, 279, 304, 365, 397, 400),
                   states = c(1, 2, 1, 3, 2, 3, 1, 2, 3, 2, 2))

set.seed(7654)
update_behav_param(const, behav_proc)
# $lambda
# [1] 0.04658404 0.05334004 0.08813678
#
# $q
# [,1]      [,2]      [,3]
# [1,]          NA 0.6686524 0.3313476
# [2,] 0.008780271        NA 0.9912197
# [3,] 0.366922353 0.6330776        NA
}
\seealso{
Other Behaviour parameters: \code{\link{calc_markov_post_param}},
  \code{\link{calc_markov_suff_stats}},
  \code{\link{check_valid_markov}}
}
\concept{Behaviour parameters}
