% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refined_path.R
\name{update_refined_path}
\alias{update_refined_path}
\title{Update a given section of the refined path.}
\usage{
update_refined_path(fixed_constant, type, fixed_values, curr_values,
  move_params, behav_params, obs)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{num_states}}{Numeric, the number of behavioural states.}
\item{\code{indep_step}}{Boolean, True if independent step model, False if correlated.}
\item{\code{obs_error}}{Boolean, True if observation error is to be included (independent or correlated).}
\item{\code{corr_obs_error}}{Boolean, True if the observation error is correlated rather than independent.}
\item{\code{q.bearings}}{Vector, index positions in \code{move_params} corresponding to bearing parameters.}
\item{\code{q.speed_mean}}{Vector, index positions in \code{move_params} corresponding to mean speed parameters.}
\item{\code{q.speed_corr}}{Vector, index positions in \code{move_params} corresponding to speed correlation parameters (only if correlated step model).}
\item{\code{q.speed_var}}{Vector, index positions in \code{move_params} corresponding to speed variance parameters.}
\item{\code{q.obs_error}}{Vector, index positions in \code{move_params} corresponding to error parameters.}
\item{\code{ideal_refined_time_diff}}{Numeric, ideal time scale for the refined path (only if multistate case).}
}}

\item{type}{Character, the type of section being updated (start, middle, or end) from within the full path.}

\item{fixed_values}{List with components set by \code{set_fixed_values}.}

\item{curr_values}{List with componenets set by \code{set_current_values}.}

\item{move_params}{Vector, movement parameters.}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector, switching rates.}
\item{\code{q}}{Square matrix (diagonal \code{NA}), switching probabilities.}
}}

\item{obs}{List with components:
\describe{
\item{\code{x}}{Vector, observed x locations.}
\item{\code{y}}{Vector, observed y locations.}
\item{\code{times}}{Vector, observation times.}
\item{\code{time_diffs}}{Vector, differences in observation times.}
\item{\code{index_on_refined_path}}{Vector, index positions on refined path of observation times.}
\item{\code{errors}}{List with components:
  \describe{
  \item{\code{x}}{Vector, errors in x direction.}
  \item{\code{y}}{Vector, errors in y direction.}
  }}
}}
}
\value{
List with the components:
  \describe{
  \item{\code{accept}}{Boolean, True if proposal is accepted.}
  \item{\code{behav_fail}}{Boolean, True if proposal failed because of the rejection step in the behavioural proposal.}
  \item{\code{bearings}}{Vector of bearings.}
  \item{\code{steps}}{Vector of steps.}
  \item{\code{times}}{Vector of times (only in multi-behaviour).}
  \item{\code{behavs}}{Vector of behaviours (only in multi-behaviour).}
  }
}
\description{
\code{update_refined_path} returns a MH update of a given section of the refined path.
}
\seealso{
Other Refined path: \code{\link{sim_steps}}
}
\concept{Refined path}
