% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{update_speed_param}
\alias{update_speed_param}
\title{Simulate speed parameters.}
\usage{
update_speed_param(fixed_constant, curr_move_params, behav_params, refined_path)
}
\arguments{
\item{fixed_constant}{List with components:
\describe{
\item{\code{pert_sd}}{Vector, perturbation standard deviations for the speed parameter proposals.}
\item{\code{num_states}}{Numeric, number of behavioural states.}
\item{\code{indep_step}}{Boolean, True if independent step model, False if correlated step model.}
\item{\code{num_speed_param}}{Numeric, number of speed parameters.}
\item{\code{q.bearings}}{Vector, index positions within \code{curr_move_params} relating to bearing parameters.}
\item{\code{q.speed_mean}}{Vector, index positions within \code{move_params} relating to speed mean parameters.}
\item{\code{q.speed_corr}}{Vector, index positions within \code{move_params} relating to speed correlation parameters.}
\item{\code{q.speed_var}}{(Will only apply if \code{indep_step} is False) Vector, index positions within \code{move_params} relating to speed variance parameters.}
\item{\code{q.obs_error}}{Vector, index positions within \code{curr_move_params} relating to observation error parameters.}
}}

\item{curr_move_params}{Vector of movement parameters (as turn_var_1, turn_var_2, mean_speed_1, .., speed_var_2).}

\item{behav_params}{List with components:
\describe{
\item{\code{lambda}}{Vector (length \code{num_states}) of switching rates.}
\item{\code{q}}{Matrix (square size \code{num_states} with \code{NA} diagonal elements) of switching probabilities. Note: row = from, column = to.}
}}

\item{refined_path}{List with components:
\describe{
\item{\code{time_diffs}}{Vector, differences in the refined time scale.}
\item{\code{behavs}}{Vector, refined behavioural/state sequence.}
\item{\code{steps}}{Vector, refined step process.}
}}
}
\value{
List with the components:
  \describe{
  \item{\code{accept_prob}}{Numeric, probability of accepting the MH proposal.}
  \item{\code{prop_speed_params}}{Vector, proposed speed parameters.}
  }
}
\description{
\code{update_speed_param} returns a conditional MH proposal (with acceptance probability) of the speed parameters.
}
\seealso{
Other Speed parameters: \code{\link{calc_cond_step_dist}},
  \code{\link{calc_markov_steady_state}},
  \code{\link{calc_step_lik}}
}
\concept{Speed parameters}
